move more routes; fix some lint errors
