+++ gdbgui/py.typed
@@ -1,0 +1,2 @@
+# Marker file for PEP 561.  This package uses inline types.
+# https://mypy.readthedocs.io/en/latest/installed_packages.html#making-pep-561-compatible-packages
